Update renderIcon()
The getIcon(...) call in this function is passed an object with both a state and a unit. The getIcon function, however, expects its argument to act directly as the key for the ICONS dictionary object. As-is getIcon(...) will return undefined. In order to fix this invocation of getIcon(...) without breaking the invocation of getIcon(...) in renderInfo(...), pass in only the state value rather than the whole object.

Further addresses #8